+++ test/intercept_sys_read.c
@@ -38,9 +38,3 @@ static __attribute__((constructor)) void init(void)
 {
     intercept_hook_point = hook;
 }
-
-int main() {
-    char buf[128];
-    read(0, buf, sizeof(buf));
-    write(1, buf, strlen(buf));  // Should print "intercepted_read"
-}
