+++ test/write_test.c
@@ -1,0 +1,7 @@
+#include <unistd.h>
+#include <string.h>
+
+int main() {
+    char buf[128] = "original_syscall\n";
+    write(1, buf, strlen(buf));
+}
